Make InstructionData derive from Hash, PartialEq, and Eq.
This allows it to be used as a key in a HashMap, without a complex
manual hashing implementation.

This entails adding Hash, PartialEq, and Eq to several utility
classes.

Notably, that includes EntityList, which cannot implement a deep
equality or hashing functions without additional context, so it just
implements shallow equality and hashing.