Simplify targets and settings for cton-util wasm.
This makes licm and simple_gvn tests require a target ISA. Those passes
don't inherently require a target ISA, though in any realistic use case
they'll have one, and this allows them to have access to the contained
Flags.

This also moves the responsibility for asking for cfg and domtree
computations the code that needs them, to reduce the number of places
that implicitly know about pass dependencies.

And it adds a more complete way to specify flags and isa to cton-wasm
via the command-line.