clif-util: schedule pass jobs incrementally while scanning directories
When scanning directories, if we were doing a non-pass test runner suite, then
scanned test files would be incrementally scheduled for execution, but if we
were doing a pass test runner suite, then they would not be. This seems to be
because the method for scheduling jobs is different for pass runs.

This commit factors out the schedule-these-jobs-please callback from scanning
directories, so that pass jobs can be incrementally scheduled and pipelined with
more directory scanning just like other jobs.