use newly-added RegUnit affinity
this affinity is set in several circumstances:
* if a value eventually used as a function argument has no other conflicting affinity
* if a FixedReg or FixedTied constraint is present on an operand
  (where the register affinity is the register the constraint indicates)
  this affects instructions like x86 shl/shr, where the shift amount is
  tied to `{r,e,}cx`
* and if an affinity is merged with a constraint that would similarly
indicate a preferred register

Reload also allows an optional register preference to reload into, which
is informed by the affinity of the value to be reloaded.

Taken all together, this addresses a case where reloads for function
arguments would go into the first available register, then if the ABI
requires a different register, copied there. Now, we load directly into
the ABI-required register, if no other affinity would prohibit that.