Move recipe insertion to encode-building time
In order to avoid ownership issues in the commits to follow, this change moves the building of the Recipes collection to EncodingBuilder::build. Prior to this, the RISCV encodings were collected in a prior step (`recipes.rs`) but now they are inserted into Recipes at the same time as x86 (`encodings.rs`). This change also moves the de-duplication logic from several places to the Recipes collection implementation.