Add CI to auto-publish to crates.io on tags
This commit updates the publication process for the cranelift suite of
crates. the `publish-all.sh` script was removed in favor of a new Rust
script located at `ci/publish.rs` This script is based off the one in
the [wasm-bindgen repository][wbg]. Additionally publication
instructions have been added to `CONTRIBUTING.md`.

The new process of publication looks like:

* Run `rustc ci/publish.rs && ./publish bump`
* Commit the change and get it merged (either push it or go through a PR)
* Push a tag with the versions bumped

The CI for the last step, the tag push, will automatically publish the
crates once CI has finished.

[wbg]: https://github.com/rustwasm/wasm-bindgen/blob/master/publish.rs