Replace legalization of x86 SIMD `ishl` with an implementation including an 8x16 legalization
Though hardly an optimal instruction to use, the `i8x16.ishl` lowering is necessary because of an insistence in the SIMD spec committee that the instruction is desirable for performance on non-x86 platforms and Wasm instruction symmetry (see https://github.com/WebAssembly/simd/issues/117). This legalization uses a 128-byte table of vectors to mask off the parts of the i8x16 vector that are unaffected by a i16x8 shift.